+++ Changelog.md
@@ -26,3 +26,4 @@
 * `Missions:crewForRent` now requires the player to have `Player:withMenu` set
 * added `Player:withQuickDial()` to display quick dials on the relay station
 * `Station:withStorage` can now handle Scan Probes as products. Use the id `scanProbe`.
+* Missions can have event listeners added from outside via `mission:addSuccessListener()`, etc.